Fix MiniExample 1: Doofuses - Gameobject version

+++ Example1-DOTS-DoofusesMustEat/DOOFUSES_GAMEOBJECTS/Packages/com.sebaslab.svelto.common/DataStructures/Arrays/FasterReadOnlyList.cs
@@ -7,7 +7,7 @@ namespace Svelto.DataStructures
         public static FasterReadOnlyList<T> DefaultEmptyList = new FasterReadOnlyList<T>(
             FasterList<T>.DefaultEmptyList);
 
-        public int count      => _list.count;
+        public int  count    => (int) _list.count;
         public uint capacity => _list.capacity;
 
         public FasterReadOnlyList(FasterList<T> list)
@@ -40,11 +40,11 @@ namespace Svelto.DataStructures
         public ref T this[uint index]
         {
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            get => ref _list[index];
+            get => ref _list[(int) index];
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public T[] ToArrayFast(out uint count)
+        public T[] ToArrayFast(out int count)
         {
             return _list.ToArrayFast(out count);
         }
