+++ Example1-DOTS-DoofusesMustEat/DOOFUSES_GAMEOBJECTS/Packages/com.sebaslab.svelto.tasks@2.0.0-preview.1/Runners/SveltoTaskRunner.cs
@@ -111,7 +111,7 @@ namespace Svelto.Tasks.Internal
 
                     if (result == false)
                     {
-                        _coroutines.UnorderedRemoveAt(index);
+                        _coroutines.UnorderedRemoveAt((uint)index);
 
                         coroutinesCount--;
                     }
