+++ Example1-DOTS-DoofusesMustEat/DOOFUSES_GAMEOBJECTS/Packages/com.sebaslab.svelto.tasks@2.0.0-preview.1/Runners/Unity/RunnerBehaviourUpdate.cs
@@ -68,7 +68,7 @@ namespace Svelto.Tasks.Unity.Internal
                     var ret = routines[i].MoveNext(profiler);
                     if (ret == false)
                     {
-                        routines.UnorderedRemoveAt(i);
+                        routines.UnorderedRemoveAt((uint)i);
                         i--;
                     }
                 }
